pyproject.toml
[tool.poetry]
name = "budgetdrfbackend"
version = "0.1.0"
description = "A backend for the Budget project."
authors = ["Elliot Redhead <elliot@elliotredhead.com>"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "budget.settings"

addopts = """
	--randomly-seed=16345
	"""

[tool.pydocstyle]
inherit = false
ignore = ["D100","D203", "D212"]

[tool.isort]
profile = "black"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
pytest-django = "^4.5.2"
pytz = "^2022.1"
python-dotenv = "^0.20.0"
django-filter = "^22.1"
drf-spectacular = "^0.22.1"
django-cors-headers = "^3.13.0"
djangorestframework-simplejwt = "^5.2.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.3.0"
Werkzeug = "^2.1.2"
pytest = "^7.1.2"
factory-boy = "^3.2.1"
Faker = "^13.13.0"
pytest-factoryboy = "^2.5.0"
snapshottest = "^0.6.0"
pytest-django = "^4.5.2"
pytest-randomly = "^3.12.0"
pydocstyle = {extras = ["toml"], version = "^6.1.1"}
django-extensions = "^3.2.0"
isort = "^5.10.1"
pre-commit = "^2.19.0"
nplusone = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
